backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
redis
requests
sqlalchemy
psycopg2-binary
supabase
